Week 3 and Week 4/enterprise-chat/backend/requirements.txt
fastapi
uvicorn
python-jose
passlib
pydantic_settings
waitress